Share the find closest index
